requirements.txt
fastapi
uvicorn
pydantic
scorecardpy
pandas
pyyaml
